requirements.txt
python-dotenv
pycryptodome